refactor: keep utility functions in dedicated folder
